Make it possible to comment on ideas
The ideas are still not able to be voted on.  I think this should be
tested with the users to see if they like it or not.

Comments are not threadable for the time being - meaning that comments
cannot be made on other comments.  This will also be saved for a later
time in the business development process.

Commenting was achieved through the "acts_as_commentable_with_threading"
gem