Add idea ranking, preferred ideas, gold status/credits
Hotness
=======
Added the hotness field to ideas as a means of sorting how the ideas
appear in the idea index.  The hotness of an idea is essentially its
ranking.  The higher the hotness, the closer to the front of the list it
will appear.

Hotness is mostly based off of how old the idea is (when it was
published), and how many upvotes it has.  This uses the same kind of
algorithm as Reddit does for sorting its posts.  You can view more
information about the algorithm at: http://amix.dk/blog/post/19588

Preferred Ideas
===============
If a user has gold status, xe can make xyr ideas have 'preferred' status,
which will give it a boost in terms of where it appears on the list.  It
will also (eventually) have a little icon associated with it,
distinguishing that it's 'preferred'.

If an idea is preferred, it's considered to have an extra 5000 upvotes in
terms of its ranking on the page.  Mathematically, this means that
a preferred idea with preferred status and no upvotes will be about the
equivalent in hotness of an idea that's two days newer with no upvotes.

Gold Status
===========
A user can attain gold status, thereby allowing xem to promote xyr ideas
to preferred.  A user's gold status lasts 7 days upon activated, and can
only be activated by paying 'gold credits'.  Once a user's gold status
runs out, xe can no longer make xyr ideas 'preferred'.

Gold Credits
============
The currency by which a user attains gold status.  One gold credit will
activate gold status for a user for 7 days.  As a promotion for launch, we
are automatically giving all users 1 gold credit.