Fix bug where voting would only work after refresh
Before, when starting at the index page and going to the idea show page.
If you tried to vote, the voting would not registers.  This is because
rails is running turbolinks and was unable to register the show page's
javascript as being loaded.

Installed the jquery-turbolinks gem to make this work.  Added it to the
gemfile, version 2.1.0, and added an include line to the js manifest to
make it work