Separate 'edit login info' to a different page
Before, all the user information was in one page, managed by devise,
making the user have to enter xyr password everytime xe wanted to change
anything, (like xyr profile picture).

This makes a new user edit page, (Governed by the users_controller), for
editing not-so-critical stuff like the user's profile picture and
username.  The critical login information forms are now accessible by
following a link on that page.

The link will go to the original devise page, which now only has forms for
changing the user's email and password.  The link for deactivating the
account is also there.