Prevent user from entering blank comments
This was done by rendering the ideas/new template in the
comments_controller in the possibility that saving the comment failed
(i.e. - the comment didn't have a body).

Granted, the syntax for this wasn't as clean as others in the fact that we
called
  flash[:alert] = "Can't create a blank comment"
Even though there's the possiblity that there might be some other error.