Remove user's ability to delete their account
This is mostly to prevent the bug from happening where a user makes
a comment, then deletes xyr account, causing the app to crash when someone
tries to enter the idea that has that comment on it.

If people can't delete their account, then there won't be any orphaned
comments.  Parents for everyone!