Send email on new ideas, comments, and subscribers
This uses the IdeaMailer class to send emails.  An email is sent to the
admins whenever a person creates a new idea.

Also, if someone comments on an idea, it will send an email to the idea
owner.

If someone subscribes to an idea, it will also send an email to the idea
owner.