Make a vote_tally column to increase performance
Before, we were constantly calling the acts_as_votable methods in order to
get the vote_tally score for items.  These were causing unneccessary hits
to the db.
This bypasses that saving the vote tally as a column in the db, so it's
locally available as soon as the idea is fetched.