Make the ability to crop the main image for ideas
This was done using the JCrop library provided at
http://deepliquid.com/content/Jcrop.html

This is almost an identical copy of what's going on in the following
Railscast:
http://railscasts.com/episodes/182-cropping-images

When the user creates or updates an idea and there is an uploaded image
present, the controller will render the idea/crop view.  This view
includes the JCrop library (for some reason, the code wouldn't work when
it wasn't included directly - also as a result, this had to be precompiled
as specified in config/initializers/assets.rb), in the script tag at the
top of the view.

When the user manipulates the crop box, the jquery fills in hidden fields
for the crop dimensions.  These dimensions are specified as attr_accessor
crop_x, crop_y, crop_w, crop_h in the idea model.  If these are present,
a callback in the model fires a reprocess method for the image and crops
the image accordingly using the custom paperclip processor at
lib/paperclip_processors/crop.rb.

WARNING: there is a bug right now to where if the uploaded image is
landscape, then it is impossible for the cropped image to take up the
entire original size of the image